// lib/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  USER: 'USER',
  VIEWER: 'VIEWER',
  EDITOR: 'EDITOR',
  MODERATOR: 'MODERATOR',
  ADMIN: 'ADMIN',
  SUPER_ADMIN: 'SUPER_ADMIN'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const AuditAction = {
  LOGIN: 'LOGIN',
  LOGOUT: 'LOGOUT',
  REGISTER: 'REGISTER',
  CREATE: 'CREATE',
  READ: 'READ',
  UPDATE: 'UPDATE',
  DELETE: 'DELETE',
  BAN: 'BAN',
  UNBAN: 'UNBAN',
  IMPERSONATE: 'IMPERSONATE',
  PASSWORD_RESET: 'PASSWORD_RESET',
  EMAIL_VERIFY: 'EMAIL_VERIFY',
  EXPORT: 'EXPORT',
  IMPORT: 'IMPORT',
  SETTINGS_UPDATE: 'SETTINGS_UPDATE',
  API_KEY_CREATE: 'API_KEY_CREATE',
  WEBHOOK_CREATE: 'WEBHOOK_CREATE',
  FILE_UPLOAD: 'FILE_UPLOAD',
  ROLE_CHANGE: 'ROLE_CHANGE',
  MONITOR_CHECK: 'MONITOR_CHECK'
} as const

export type AuditAction = (typeof AuditAction)[keyof typeof AuditAction]


export const PostStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
  ARCHIVED: 'ARCHIVED'
} as const

export type PostStatus = (typeof PostStatus)[keyof typeof PostStatus]